Refactored ScheduledEventBot and TriggeredBot
