+++ webofneeds/won-bot/src/main/java/won/bot/app/BotRunnerApp.java
@@ -18,7 +18,7 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.scheduling.support.PeriodicTrigger;
 
 import won.bot.framework.bot.Bot;
-import won.bot.framework.bot.base.TriggeredBot;
+import won.bot.framework.bot.base.ScheduledTriggerBot;
 import won.bot.framework.manager.BotManager;
 
 /**
@@ -42,10 +42,10 @@ public class BotRunnerApp {
         bot = (Bot) botBean;
         // the bot also needs a trigger so its act() method is called regularly.
         // (there is no trigger bean in the context)
-        if (bot instanceof TriggeredBot) {
+        if (bot instanceof ScheduledTriggerBot) {
             PeriodicTrigger trigger = new PeriodicTrigger(5000, TimeUnit.MILLISECONDS);
             trigger.setInitialDelay(1000);
-            ((TriggeredBot) bot).setTrigger(trigger);
+            ((ScheduledTriggerBot) bot).setTrigger(trigger);
         }
         BotManager botManager = (BotManager) applicationContext.getBean("botManager");
         // adding the bot to the bot manager will cause it to be initialized.
