+++ webofneeds/won-bot/src/main/java/won/bot/framework/bot/base/BaseBot.java
@@ -87,10 +87,8 @@ public abstract class BaseBot implements Bot {
     }
 
     /**
-     * Override this method to do initialization work.
+     * Override this method to add additional initialization routines.
      */
-    protected abstract void doInitialize();
-
     @Override
     public synchronized void initialize() throws Exception {
         if (!this.lifecyclePhase.isDown())
@@ -105,21 +103,17 @@ public abstract class BaseBot implements Bot {
             logger.error("Bot cannot establish connection with bot context");
             throw e;
         }
-        doInitialize();
         this.lifecyclePhase = BotLifecyclePhase.ACTIVE;
     }
 
     /**
-     * Override this method to do free resources during shutdown.
+     * Override this method to add additional shutdown routines.
      */
-    protected abstract void doShutdown();
-
     @Override
     public synchronized void shutdown() throws Exception {
         if (!this.lifecyclePhase.isActive())
             return;
         this.lifecyclePhase = BotLifecyclePhase.SHUTTING_DOWN;
-        doShutdown();
         this.lifecyclePhase = BotLifecyclePhase.DOWN;
     }
 
@@ -133,14 +127,14 @@ public abstract class BaseBot implements Bot {
         return nodeURISource;
     }
 
-    protected MatcherNodeURISource getMatcheNodeURISource() {
-        return matcherNodeURISource;
-    }
-
     @Qualifier("default")
     @Autowired(required = true)
     public void setNodeURISource(final NodeURISource nodeURISource) {
         this.nodeURISource = nodeURISource;
+    }
+
+    protected MatcherNodeURISource getMatcherNodeURISource() {
+        return matcherNodeURISource;
     }
 
     @Qualifier("default")
@@ -166,7 +160,7 @@ public abstract class BaseBot implements Bot {
     @Qualifier("default")
     @Autowired(required = true)
     public void setMatcherProtocolAtomServiceClient(
-                    final MatcherProtocolAtomServiceClientSide matcherProtocolAtomServiceClient) {
+            final MatcherProtocolAtomServiceClientSide matcherProtocolAtomServiceClient) {
         this.matcherProtocolAtomServiceClient = matcherProtocolAtomServiceClient;
     }
 
@@ -177,7 +171,7 @@ public abstract class BaseBot implements Bot {
     @Qualifier("default")
     @Autowired(required = true)
     public void setMatcherProtocolMatcherService(
-                    final MatcherProtocolMatcherServiceImplJMSBased matcherProtocolMatcherService) {
+            final MatcherProtocolMatcherServiceImplJMSBased matcherProtocolMatcherService) {
         this.matcherProtocolMatcherService = matcherProtocolMatcherService;
     }
 
@@ -225,7 +219,7 @@ public abstract class BaseBot implements Bot {
 
     @Override
     public abstract void onNewAtomCreated(final URI atomUri, final URI wonNodeUri, final Dataset atomDataset)
-                    throws Exception;
+            throws Exception;
 
     @Override
     public abstract void onConnectFromOtherAtom(Connection con, final WonMessage wonMessage);
@@ -256,7 +250,7 @@ public abstract class BaseBot implements Bot {
 
     @Override
     public abstract void onNewAtomCreatedNotificationForMatcher(final URI wonNodeURI, final URI atomURI,
-                    final Dataset atomDataset);
+            final Dataset atomDataset);
 
     @Override
     public abstract void onAtomActivatedNotificationForMatcher(final URI wonNodeURI, final URI atomURI);
