+++ webofneeds/won-bot/src/main/java/won/bot/framework/bot/base/FactoryBot.java
@@ -36,7 +36,7 @@ public abstract class FactoryBot extends EventBot {
             }
         };
         factoryBotInitBehaviour.onDeactivateActivate(runningBehaviour, factoryBotHintBehaviour,
-                        messageCommandBehaviour);
+                messageCommandBehaviour);
         factoryBotInitBehaviour.activate();
     }
